Updated Occlusion Queries to v3.3.0

+++ src/BabylonCpp/include/babylon/core/hash.h
@@ -38,7 +38,7 @@ constexpr HashValue kHashOffsetBasis     = 0x84222325;
 constexpr HashValue kHashPrimeMultiplier = 0x000001b3;
 
 BABYLON_SHARED_EXPORT HashValue Hash(const char* str);
-BABYLON_SHARED_EXPORT HashValue Hash(const char* str, size_t len);
+BABYLON_SHARED_EXPORT HashValue Hash(const char* str, int len);
 BABYLON_SHARED_EXPORT HashValue Hash(string_view str);
 BABYLON_SHARED_EXPORT HashValue HashCaseInsensitive(const char* str,
                                                     size_t len);
